pyproject.toml
[project]
name = "netcore-generator"
version = "0.1.0"
description = "A utility to generate .NET Core 9 Clean Architecture applications from PostgreSQL schemas"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "psycopg2-binary>=2.9.9",
    "jinja2>=3.1.3",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
]

